AWE-19 refactor: Reducir duplicidad
Se estaba duplicando un método para crear productos.
