+++ cart/tests.py
@@ -6,6 +6,7 @@ from django.test import TestCase
 
 from cart.cart import Cart
 from base import models as base_models
+from acme_wedding.test_utils import create_product
 import datetime
 import time
 
@@ -17,27 +18,6 @@ from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.common.keys import Keys
 
 IMAGE_PATH = 'media/agotado.png'
-
-def create_product():
-    test_category = base_models.Category()
-    test_category.name = 'Test Category'
-    test_category.slug = 'test-category'
-    test_category.save()
-
-    test_product = base_models.Product()
-    test_product.amount = 10
-    test_product.available = True
-    test_product.name = 'Producto Test'
-    test_product.slug = 'producto-test'
-    test_product.description = 'Producto de prueba para la aplicacion'
-    test_product.image = 'ImagenDePruebaXiquillo'
-    test_product.price = 100
-    test_product.created = datetime.datetime.now()
-    test_product.updated = datetime.datetime.now()
-    test_product.category = test_category
-    test_product.save()
-
-    return [test_category,test_product]
 
 class CartTest(TestCase):
     def setUp(self):
